pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "paper-edge-detection"
requires-python = ">=3.12"

dependencies = [
  "roboflow==1.1.61",
  "ultralytics==8.3.111",
  "python-dotenv (>=1.1.0,<2.0.0)"
]

